tasks/misc/WorldSudokuChampionship/deploy/requirements.txt
pyfiglet
loguru
py-sudoku
